refactor: consolidate test helper methods into own file
